--- common/include/Types.h
@@ -1,20 +1,0 @@
-
-#ifndef COMMON_TYPES_H
-#define COMMON_TYPES_H
-
-    namespace Types {
-        typedef unsigned char		UInt8;		//!< An unsigned 8-bit integer value
-        typedef unsigned short		UInt16;		//!< An unsigned 16-bit integer value
-        typedef unsigned long		UInt32;		//!< An unsigned 32-bit integer value
-        typedef unsigned long long	UInt64;		//!< An unsigned 64-bit integer value
-        typedef signed char			SInt8;		//!< A signed 8-bit integer value
-        typedef signed short		SInt16;		//!< A signed 16-bit integer value
-        typedef signed long			SInt32;		//!< A signed 32-bit integer value
-        typedef signed long long	SInt64;		//!< A signed 64-bit integer value
-        typedef float				Float32;	//!< A 32-bit floating point value
-        typedef double				Float64;	//!< A 64-bit floating point value
-
-    }
-
-
-#endif //F4MPREVERSE_TYPES_H
