--- common/include/main.h
@@ -1,1 +1,0 @@
-#pragma once
