+++ src/DirectXHook.h
@@ -11,6 +11,8 @@
 #include <imgui_impl_dx11.h>
 #include <imgui_impl_win32.h>
 
+#include "Hooks.h"
+
 using namespace Exceptions::Core::Exceptions;
 using namespace Hooks;
